+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRTZZ3D0YN8C0TP7NFF3XKK2"
+        "01KRWQHHWD7W74ENW8TXSNR6ZN"
     )
     r"""The unique identifier of the data source"""