+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRTZZ3CXGM7V2J85C7QTGJ8V"
+        "01KRWQHHWB1FKMBZBB4FP3AG6E"
     )
     r"""The unique identifier of the data source"""