+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRTZZ3CZZ2WV5Y92BQAGNG5N"
+        "01KRWQHHWD7KYZW2W3Y8Q2G4C8"
     )
     r"""The unique identifier of the data source"""