+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRTZZ3D2SJFNAK6BV8JBBRKY"
+        "01KRWQHHWFH3QAMWPY9P0ZWZM7"
     )
     r"""The unique identifier of the data source"""