+++ packages/orq-rc/RELEASES.md
@@ -7969,3 +7969,13 @@ Based on:
 - [python v4.10.0-rc.10] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.10 - packages/orq-rc
+
+## 2026-05-18 05:07:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.11 - packages/orq-rc